refactor: rename-directory
